Generate truststore jks file when generating Certificate files
A java 'truststore' keystore is a jks format file that contains
the CA cert(s) that should be used to verify client certificates.

If Java services are configured with a truststore, they will
only be allowed to communicate over TLS/SSL with clients
that use certificates that are signed by the CA in the truststore.

Bug: T166167
Change-Id: I92f3bd6a73f3f506b5d92cfbdb06cf50ffb6cfc2